Upd: improve docs, split module `Common`
