+++ dom-parser.cabal
@@ -57,8 +57,9 @@ library
                      , xml-conduit
                      , xml-lens
   exposed-modules: Text.XML.DOM.Parser
+                 , Text.XML.DOM.Parser.Buildable
                  , Text.XML.DOM.Parser.Combinators
-                 , Text.XML.DOM.Parser.Common
+                 , Text.XML.DOM.Parser.Content
                  , Text.XML.DOM.Parser.FromAttribute
                  , Text.XML.DOM.Parser.FromDom
                  , Text.XML.DOM.Parser.Types
