+++ src/Text/XML/DOM/Parser.hs
@@ -1,13 +1,15 @@
 module Text.XML.DOM.Parser
-  ( module Text.XML.DOM.Parser.Combinators
-  , module Text.XML.DOM.Parser.Common
+  ( module Text.XML.DOM.Parser.Buildable
+  , module Text.XML.DOM.Parser.Combinators
+  , module Text.XML.DOM.Parser.Content
   , module Text.XML.DOM.Parser.FromAttribute
   , module Text.XML.DOM.Parser.FromDom
   , module Text.XML.DOM.Parser.Types
   ) where
 
+import Text.XML.DOM.Parser.Buildable
 import Text.XML.DOM.Parser.Combinators
-import Text.XML.DOM.Parser.Common
+import Text.XML.DOM.Parser.Content
 import Text.XML.DOM.Parser.FromAttribute
 import Text.XML.DOM.Parser.FromDom
 import Text.XML.DOM.Parser.Types
