+++ src/Text/XML/DOM/Parser/FromDom.hs
@@ -27,7 +27,7 @@ import Data.Text as T hiding (empty)
 import Data.Typeable
 import Data.Void
 import Text.XML
-import Text.XML.DOM.Parser.Common
+import Text.XML.DOM.Parser.Content
 import Text.XML.DOM.Parser.Types
 import TypeFun.Data.List hiding (Union)
